// PRO-EP/pro-ep-ui/src/app/interfaces/interfaces.ts
export interface IUser {
  userid: number;
  dempoid: string;
  status: string;
  active?: boolean | null;
  corehours?: number | null;
  employmenttype?: number | null;
  fname: string;
  lname: string;
  permstartdate?: Date | null;
  permenddate?: Date | null;
  tempstartdate?: Date | null;
  tempenddate?: Date | null;
  uinit: string;
  role?: number | null;
  schedulinglevel?: number | null;
  scheduledisplay?: number | null;
  title: string;
  defaultproject?: number | null;
  phonenumber: string;
  phonenumber1: string;
  phonenumber2: string;
  emailaddr: string;
  emailaddr1: string;
  bypasslockout?: number | null;
  aislenumber?: number | null;
  teamgroup: string;
  workgroup: string;
  cubeoffice?: number | null;
  spanish?: number | null;
  language: string;
  canedit?: boolean | null;
  trainedon: string;
  citidate?: Date | null;
  introemail?: boolean | null;
  introemaildate?: Date | null;
  phonescreen?: boolean | null;
  phonescreendate?: Date | null;
  facetoface?: boolean | null;
  facetofacedate?: Date | null;
  welcomeemail?: boolean | null;
  welcomeemaildate?: Date | null;
  orientationdate?: Date | null;
  referral: string;
  tempagency: string;
  userImage: string | null;
  notes: string;
  emercontactnumber: string;
  emercontactnumber2: string;
  emercontactname: string;
  emercontactname2: string;
  emercontactrel: string;
  emercontactrel2: string;
  dob?: Date | null;
  state: string;
  city: string;
  zipcode: string;
  homeaddress: string;
  phonenumber3: string;
  hiatusstartdate?: Date | null;
  computernumber: string;
  badgeidnumber: string;
  acdagentnumber: string;
  uniqueid: string;
  preferredfname: string;
  preferredlname: string;
  entryFormName: string;
  buddy?: boolean;
  entryDt?: Date | null;
  entryBy: string;
  modDt?: Date | null;
  modBy: string;
}
// ...
export interface ITimeCard {
  timeCardId: number;
  dempoid: string;
  //workDate?: Date;
  timeIn?: Date | null;
  timeOut?: Date | null;
  entryBy: string;
  entryDt?: Date;
  modBy: string;
  modDt?: Date;
  datetimein?: Date | null;
  datetimeout?: Date | null;
  externalIp: string | null;
  //MachineName: string;
}

export interface ICurrentUser {
  user: IUser;
  timecards: ITimeCard[];
}
// ...
export interface IAuthenticatedUser {
  displayName: string;
  duDukeID: string;
  eppn: string;
  netID: string;
  isMemberOf: string;
  interviewer: boolean;
  resourceGroup: boolean;
  admin: boolean;
  userRoles?: string[] | undefined;
  sessionMinsLeft: number;
  timecards: ITimeCard[];
}

export interface IDateRange {
  startDate: Date;
  endDate: Date;
}


export interface IUserTrainedOn {
  netId: string;
  trainedOn: string[];
}

export interface IPopupMessage {
  title: string;
  body: string;
  buttons: IActionButton[] | null;
  closeCallback: Function | null;
  noBlackFilter?: boolean;
}

export interface IActionButton {
  label: string;
  callbackFunction?: Function;
}